Postgres config

+++ db/migrate/20230112084653_create_customers.rb
@@ -3,7 +3,7 @@ class CreateCustomers < ActiveRecord::Migration[7.0]
     create_table :customers do |t|
       t.string :name
       t.string :email
-      t.integer :phone_number
+      t.numeric :phone_number
       t.string :photo_url
       t.string :location
       t.string :password_digest
